requirements.txt
python-binance==1.0.19
pandas==2.1.4
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.10
langchain-together==0.0.1
langgraph==0.0.20
pydantic==2.5.3
streamlit==1.31.0
plotly==5.18.0